Improvements in sporadic checks
In sporadic_checks.m and shimura_curve_CM_locus, we now use the best known progress on Selberg's eigenvalue conjecture due to Kim--Sarnak to improve the version of Abramovich's lower bound on the gonality of X_0^D(N) we use. We update our lists pertaining to sporadic point existence accordingly.